Revert "BDOG-3046 separate route for the callback endpoint"
