+++ conf/app.routes
@@ -1,12 +1,16 @@
 # microservice specific routes
 GET        /assets/*file                       controllers.Assets.versioned(path = "/public", file: Asset)
-GET        /hello-world                        uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.show()
-GET        /v2/hello-world                     uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showV2()
+GET        /hello-world                        uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.show
+GET        /v2/hello-world                     uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showV2
 
 GET        /hello-world/submission-form        uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showSubmissionForm(uploadId: uk.gov.hmrc.helloworldupscan.model.UploadId)
-POST       /hello-world/submission-form        uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.submitFormWithFile()
-GET        /hello-world/submission-form/result uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showSubmissionResult()
+POST       /hello-world/submission-form        uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.submitFormWithFile
+GET        /hello-world/submission-form/result uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showSubmissionResult
 
 # upload file redirects
 GET        /hello-world/success                uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showResult(uploadId: uk.gov.hmrc.helloworldupscan.model.UploadId)
 GET        /hello-world/error                  uk.gov.hmrc.helloworldupscan.controllers.UploadFormController.showError(errorCode: String, errorMessage: String, errorRequestId: String, key: String)
+
+# upload file notification callback
++ nocsrf
+POST       /hello-world/upscan-callback        uk.gov.hmrc.helloworldupscan.controllers.UploadCallbackController.callback
