+++ conf/application.conf
@@ -53,11 +53,7 @@ microservice {
 upload-redirect-target-base = "http://localhost:9000"
 
 upscan {
-  # In a deployed environment, it's important that the callback url that is sent
-  # to upscan-initiate should be the internal host pattern e.g. my-service.protected.mdtp
-  # This is because upscan-initiate does not send the callback over the public internet,
-  # your callback endpoint should not be accessible via the public internet.
-  callback-endpoint = "http://localhost:9000/internal/hello-world/upscan-callback"
+  callback-endpoint = "http://localhost:9000/hello-world-upscan/hello-world/upscan-callback"
 }
 
 google-analytics {
