+++ conf/prod.routes
@@ -1,4 +1,3 @@
 # Add all the application routes to the app.routes file
 ->         /hello-world-upscan        app.Routes
-->         /internal                  internal.Routes
 ->         /                          health.Routes
