+++ test/uk/gov/hmrc/helloworldupscan/controllers/CallbackBodyTest.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.helloworldupscan.controllers.internal
+package uk.gov.hmrc.helloworldupscan.controllers
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
